changed to use real position, not integer
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@8168 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/focusd/fli.cpp
@@ -46,7 +46,7 @@ class Fli:public Focusd
 		virtual int init ();
 		virtual int initValues ();
 		virtual int info ();
-		virtual int setTo (int num);
+		virtual int setTo (float num);
 		Rts2ValueLong *focExtent;
 
 	public:
@@ -200,7 +200,7 @@ int Fli::info ()
 	return Focusd::info ();
 }
 
-int Fli::setTo (int num)
+int Fli::setTo (float num)
 {
 	LIBFLIAPI ret;
 	ret = info ();
